clean up old code
