+++ esp32/commands/at_command.c
@@ -93,51 +93,6 @@ void esp32_handle_line(ESP32 *inst, const char *line) {
     }
 }
 
-void esp32_handle_ipd(struct esp32state *inst, const char *line) {
-    //receivedBytesRemaining
-    char *token = strtok(line, ",");
-    int linkId = atoi(strtok(NULL, ","));
-    int length = atoi(strtok(NULL, ":"));
-
-    xil_printf("[ IPD] socket data received, link: %d, length: %d\r\n", linkId, length);
-    struct at_socket * socket = &inst->sockets[linkId];
-
-    // TODO: handle this more elegantly, prob by capturing this in the ISR
-    // copy beginning bits
-    //char *begin = strtok(NULL, "\0");
-    //int beginLength = strlen(begin);
-
-    //memcpy(socket->buffer, begin, beginLength);
-    // this is shit
-    //socket->buffer[beginLength] = '\r';
-    //socket->buffer[beginLength + 1] = '\n';
-
-    //socket->buffer_size = beginLength + 2;
-    socket->receivedBytesRemaining = length;
-
-    inst->active_socket = socket;
-    esp32_set_receive(inst, &esp32_handle_socket_recv);
-}
-
-void esp32_handle_socket_recv(struct esp32state *inst, char c) {
-    /* +IPD,<link_id>,<len> */
-    struct at_socket * socket = inst->active_socket;
-
-    socket->buffer[socket->buffer_size++] = c;
-    //TODO check for overflow
-
-    if (--socket->receivedBytesRemaining == 0) {
-        socket->buffer[socket->buffer_size] = '\0';
-        esp32_set_receive(inst, &esp32_on_receive);
-
-        //xil_printf("received all data: %s\r\n", socket->buffer);
-
-    } else if (socket->receivedBytesRemaining < 50) {
-        //xil_printf("%d.", socket->receivedBytesRemaining);
-    }
-
-}
-
 
 /**
  * A function to handle an unexpected command response
