+++ esp32/commands/at_command.h
@@ -15,6 +15,5 @@ struct at_command* esp32_lookup_command(struct esp32state *inst, const char *lin
 bool esp32_is_response_end(const char *line);
 void esp32_handle_line(struct esp32state *inst, const char *line);
 void esp32_handle_unexpected(struct esp32state *inst, const char *line);
-void esp32_handle_socket_recv(struct esp32state *inst, char c);
 
 #endif
